// src/pages/CustomerManagement/customerDetail/dxl/data.d.ts
export interface FollowData {
  showFollowButton: number;
  showFollowProductAndCoupon: number;
  showDistributedCompany: 0 | 1;  // 是否显示派发公司
  distributedCompany: any[];  // 派发公司列表
  followTab: FollowTab[];
}

export interface FollowTab {
  key: string;
  val: string;
}
// ...
export interface CustomerData {
  customerId: string;
  customerLevel: string;
  customerLevelText: string;
  channel: string;
  customerName: string;
  gender: string;
  genderText: string;
  identity: string;
  identityText: string;
  weddingDate: string;
  phone: string;
  weChat: string;
  liveCityInfo: CityInfo;
  liveAddress: string;
  workCityInfo: CityInfo;
  workAddress: string;
  contactTime: string;
  referrerName: string;
  likeCityInfo: CityInfo;
  budget: string;
  category: string;
  weddingStyle: string;
  weddingStyleText: string;
  deskNum: string;
  hotel: string;
  encryptPhone: string;
  nextContactTime: string;
  allotTime: string;
  createTime: string;
  createUser: string;

  referrerPhone: string;
  customerFollowStatus: string;
  editable: 0 | 1; // 0 不可编辑  1 可编辑

  status_txt: string;

  repeatStatus: 1 | 2 | 3;  //  客户重单状态 1:正常,2:客户合并,3:亲子单
  repeatAuditStatus: 0 | 1 | 2 | 3;  //  客户重单审核状态 0:正常无重单提审 1:待审核,2:驳回,3:通过
  status: string;   // 客户状态
  complaint_status: 0 | 1 | 2 | 3;  // 客诉审核状态
  
  weddingDateFrom:string;
  weddingDateEnd:string;

  allot_time:string; //客户划入时间
  record_user_id: string;  //(北京)提供人ID
  record_user_name: string; //(北京)提供人
}
// ...
export interface CityInfo {
  province; string;
  city: string;
  district: string;
  full: string;
  city_code: string;
}
// ...
export interface FollowListItem {
  followTime: string
  callDuration: string
  followType: string
  results: string
  comment: string
  nextContactTime: string
  contactWay: string
  followUser: string
  attachment: array
  followTag: string
  arrivalTime: string
  couponName: string
  productName: string
  state: string
}